Separate the ferry connection logic and the node expansion into separate files.
